created first controller
