+++ routes.js
@@ -1,24 +1,9 @@
 const express = require('express')
 const router = express.Router()
 
-const db = require ('./conn.js')
+const listingController = require("./listingController")
 
 router.route("/listings")
-    .get(async function (_req, res){
-        // res.send("success!")
-        const dbConn = db.getDbConn()
-        try {
-            console.log("fetching listings")
-            const listings = await dbConn
-                                        .collection('listingsAndReviews')
-                                        .find({})
-                                        .limit(50)
-                                        .toArray()
-            res.status(200).json(listings)
-        } catch (error){
-            console.error(error)
-            res.status(400)
-        }
-    })
+    .get(listingController.index)
 
 module.exports = router;
